refactor(state): clarify prefill checkpoint module name
