+++ python/tokenspeed/runtime/layers/attention/backends/state/mamba.py
@@ -41,7 +41,7 @@ from tokenspeed_kernel.ops.attention.gdn.triton import (
     set_total_chunks_hint_uniform,
 )
 from tokenspeed_kernel.ops.attention.kda.triton import verify_state_blocks
-from tokenspeed_kernel.ops.attention.triton.prefill_checkpoints import (
+from tokenspeed_kernel.ops.attention.triton.prefill_state_checkpoints import (
     pack_prefill_recurrent_checkpoint_inputs,
     write_prefill_conv_checkpoints,
     write_prefill_recurrent_checkpoints,
